tests/wpt/tests/subapps/insecure-context-error.tentative.html

<!DOCTYPE html>
<title>Sub Apps: subApps is undefined in insecure context (non-https)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

test(() => {
  assert_equals(window.subApps, undefined);
  assert_equals(window.SubApps, undefined);
}, 'subApps is not defined in insecure context.');

</script>
